Shift votefor&delete answer views&shift tests
  1. Shift 'votefor' view from user app to question app
  2. Delete answer views (create, delete). Now deleting and creating answer
     only in update question way.
  3. Also shift test&url for 'votefor' view to question app
