+++ questions/urls.py
@@ -4,14 +4,11 @@ from django.conf.urls import url
 from .views import (QuestionList,
                     QuestionCreate,
                     QuestionRetrieveUpdateDestroy,
-                    AnswerDestroy,
-                    AnswerCreate)
+                    votefor)
 
 urlpatterns = [
-    url(r'^(?P<pk>[0-9]+)/$', QuestionRetrieveUpdateDestroy.as_view()),
+    url(r'^(?P<quest>[0-9]+)/votefor/(?P<pk>[0-9]+)/$', votefor),
     url(r'^create/$', QuestionCreate.as_view()),
-    url(r'^(?P<quest>[0-9]+)/answer/(?P<pk>[0-9]+)/$', AnswerDestroy.as_view()),
-    url(r'^(?P<pk>[0-9]+)/answer/create/$', AnswerCreate.as_view()),
-    url(r'^(?P<pk>[0-9]+)/$', QuestionRetrieveUpdateDestroy.as_view()),
     url(r'^all/$', QuestionList.as_view()),
+    url(r'^(?P<pk>[0-9]+)/$', QuestionRetrieveUpdateDestroy.as_view()),
 ]
