+++ questions/views.py
@@ -1,18 +1,19 @@
 """Views for questions' app."""
 
 from django.contrib.auth.models import User
+from django.core.exceptions import ObjectDoesNotExist
 
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from rest_framework.generics import (ListAPIView,
                                      RetrieveUpdateDestroyAPIView,
-                                     DestroyAPIView,
                                      CreateAPIView,
                                      )
 
 from .models import Question, Answer
-from .serializers import QuestionSerializer, AnswerSerializer
+from .serializers import QuestionSerializer
 from .permissions import IsOwnerOrReadOnly
 
 
@@ -50,29 +51,27 @@ class QuestionRetrieveUpdateDestroy(RetrieveUpdateDestroyAPIView):
         return super(QuestionRetrieveUpdateDestroy, self).update(request, **kwargs)
 
 
-class AnswerDestroy(DestroyAPIView):
-    """Delete API view for Answer model."""
+@api_view(['POST', ])
+@permission_classes((IsAuthenticated, ))
+def votefor(request, quest, pk):
+    """Create a vote to the corresponding answer."""
+    user = request.user
+    try:
+        answer = Answer.objects.get(id=pk)
+        question = Question.objects.get(id=quest)
+    except ObjectDoesNotExist:
+        return Response({'status': 'failed',
+                         'detail': 'doesn\'t exist given answer or question'},
+                        status=status.HTTP_404_NOT_FOUND)
 
-    serializer_class = AnswerSerializer
-    queryset = Answer.objects.all()
+    if question not in user.userprofile.voted_posts.all():
+        user.userprofile.voted_posts.add(question)
+        answer.votes_count += 1
+        answer.save()
+    else:
+        return Response({'status': 'failed',
+                         'detail': 'already voted'},
+                        status=status.HTTP_400_BAD_REQUEST)
 
-
-class AnswerCreate(CreateAPIView):
-    """Create API view for Answer model."""
-
-    serializer_class = AnswerSerializer
-
-    def create(self, request, pk):
-        """Create answer with given data for question(id=pk)."""
-        question = Question.objects.get(id=pk)
-        answer_data = request.data.get('answer')
-
-        if len(answer_data) == 0:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-
-        answer, created = Answer.objects.get_or_create(question=question, answer=answer_data)
-        serializer = self.serializer_class(answer)
-        if created:
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        else:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+    return Response({'status': 'sucess'},
+                    status=status.HTTP_202_ACCEPTED)
